Move the round functionality and its tests into a separate module.
