+++ cpp/src/arrow/CMakeLists.txt
@@ -438,6 +438,7 @@ if(ARROW_COMPUTE)
        compute/kernels/scalar_if_else.cc
        compute/kernels/scalar_nested.cc
        compute/kernels/scalar_random.cc
+       compute/kernels/scalar_round.cc
        compute/kernels/scalar_set_lookup.cc
        compute/kernels/scalar_string_ascii.cc
        compute/kernels/scalar_string_utf8.cc
