+++ cpp/src/arrow/compute/kernels/scalar_arithmetic.cc
@@ -56,32 +56,6 @@ using applicator::ScalarUnaryNotNullStateful;
 
 namespace {
 
-// Convenience visitor to detect if a numeric Scalar is positive.
-struct IsPositiveVisitor {
-  bool result = false;
-
-  template <typename... Ts>
-  Status Visit(const NumericScalar<Ts...>& scalar) {
-    result = scalar.value > 0;
-    return Status::OK();
-  }
-  template <typename... Ts>
-  Status Visit(const DecimalScalar<Ts...>& scalar) {
-    result = scalar.value > 0;
-    return Status::OK();
-  }
-  Status Visit(const Scalar& scalar) { return Status::OK(); }
-};
-
-bool IsPositive(const Scalar& scalar) {
-  IsPositiveVisitor visitor{};
-  std::ignore = VisitScalarInline(scalar, &visitor);
-  return visitor.result;
-}
-
-// N.B. take care not to conflict with type_traits.h as that can cause surprises in a
-// unity build
-
 // Bitwise operations
 
 struct BitWiseNot {
@@ -449,556 +423,6 @@ struct LogbChecked {
       return x;
     }
     return std::log(x) / std::log(base);
-  }
-};
-
-struct RoundUtil {
-  // Calculate powers of ten with arbitrary integer exponent
-  template <typename T = double>
-  static enable_if_floating_value<T> Pow10(int64_t power) {
-    static constexpr T lut[] = {1e0F, 1e1F, 1e2F,  1e3F,  1e4F,  1e5F,  1e6F,  1e7F,
-                                1e8F, 1e9F, 1e10F, 1e11F, 1e12F, 1e13F, 1e14F, 1e15F};
-    int64_t lut_size = (sizeof(lut) / sizeof(*lut));
-    int64_t abs_power = std::abs(power);
-    auto pow10 = lut[std::min(abs_power, lut_size - 1)];
-    while (abs_power-- >= lut_size) {
-      pow10 *= 1e1F;
-    }
-    return (power >= 0) ? pow10 : (1 / pow10);
-  }
-};
-
-// Specializations of rounding implementations for round kernels
-template <typename Type, RoundMode>
-struct RoundImpl;
-
-template <typename Type>
-struct RoundImpl<Type, RoundMode::DOWN> {
-  template <typename T = Type>
-  static constexpr enable_if_floating_value<T> Round(const T val) {
-    return std::floor(val);
-  }
-
-  template <typename T = Type>
-  static enable_if_decimal_value<T, void> Round(T* val, const T& remainder,
-                                                const T& pow10, const int32_t scale) {
-    (*val) -= remainder;
-    if (remainder.Sign() < 0) {
-      (*val) -= pow10;
-    }
-  }
-};
-
-template <typename Type>
-struct RoundImpl<Type, RoundMode::UP> {
-  template <typename T = Type>
-  static constexpr enable_if_floating_value<T> Round(const T val) {
-    return std::ceil(val);
-  }
-
-  template <typename T = Type>
-  static enable_if_decimal_value<T, void> Round(T* val, const T& remainder,
-                                                const T& pow10, const int32_t scale) {
-    (*val) -= remainder;
-    if (remainder.Sign() > 0 && remainder != 0) {
-      (*val) += pow10;
-    }
-  }
-};
-
-template <typename Type>
-struct RoundImpl<Type, RoundMode::TOWARDS_ZERO> {
-  template <typename T = Type>
-  static constexpr enable_if_floating_value<T> Round(const T val) {
-    return std::trunc(val);
-  }
-
-  template <typename T = Type>
-  static enable_if_decimal_value<T, void> Round(T* val, const T& remainder,
-                                                const T& pow10, const int32_t scale) {
-    (*val) -= remainder;
-  }
-};
-
-template <typename Type>
-struct RoundImpl<Type, RoundMode::TOWARDS_INFINITY> {
-  template <typename T = Type>
-  static constexpr enable_if_floating_value<T> Round(const T val) {
-    return std::signbit(val) ? std::floor(val) : std::ceil(val);
-  }
-
-  template <typename T = Type>
-  static enable_if_decimal_value<T, void> Round(T* val, const T& remainder,
-                                                const T& pow10, const int32_t scale) {
-    (*val) -= remainder;
-    if (remainder.Sign() < 0) {
-      (*val) -= pow10;
-    } else if (remainder.Sign() > 0 && remainder != 0) {
-      (*val) += pow10;
-    }
-  }
-};
-
-// NOTE: RoundImpl variants for the HALF_* rounding modes are only
-// invoked when the fractional part is equal to 0.5 (std::round is invoked
-// otherwise).
-
-template <typename Type>
-struct RoundImpl<Type, RoundMode::HALF_DOWN> {
-  template <typename T = Type>
-  static constexpr enable_if_floating_value<T> Round(const T val) {
-    return RoundImpl<T, RoundMode::DOWN>::Round(val);
-  }
-
-  template <typename T = Type>
-  static enable_if_decimal_value<T, void> Round(T* val, const T& remainder,
-                                                const T& pow10, const int32_t scale) {
-    RoundImpl<T, RoundMode::DOWN>::Round(val, remainder, pow10, scale);
-  }
-};
-
-template <typename Type>
-struct RoundImpl<Type, RoundMode::HALF_UP> {
-  template <typename T = Type>
-  static constexpr enable_if_floating_value<T> Round(const T val) {
-    return RoundImpl<T, RoundMode::UP>::Round(val);
-  }
-
-  template <typename T = Type>
-  static enable_if_decimal_value<T, void> Round(T* val, const T& remainder,
-                                                const T& pow10, const int32_t scale) {
-    RoundImpl<T, RoundMode::UP>::Round(val, remainder, pow10, scale);
-  }
-};
-
-template <typename Type>
-struct RoundImpl<Type, RoundMode::HALF_TOWARDS_ZERO> {
-  template <typename T = Type>
-  static constexpr enable_if_floating_value<T> Round(const T val) {
-    return RoundImpl<T, RoundMode::TOWARDS_ZERO>::Round(val);
-  }
-
-  template <typename T = Type>
-  static enable_if_decimal_value<T, void> Round(T* val, const T& remainder,
-                                                const T& pow10, const int32_t scale) {
-    RoundImpl<T, RoundMode::TOWARDS_ZERO>::Round(val, remainder, pow10, scale);
-  }
-};
-
-template <typename Type>
-struct RoundImpl<Type, RoundMode::HALF_TOWARDS_INFINITY> {
-  template <typename T = Type>
-  static constexpr enable_if_floating_value<T> Round(const T val) {
-    return RoundImpl<T, RoundMode::TOWARDS_INFINITY>::Round(val);
-  }
-
-  template <typename T = Type>
-  static enable_if_decimal_value<T, void> Round(T* val, const T& remainder,
-                                                const T& pow10, const int32_t scale) {
-    RoundImpl<T, RoundMode::TOWARDS_INFINITY>::Round(val, remainder, pow10, scale);
-  }
-};
-
-template <typename Type>
-struct RoundImpl<Type, RoundMode::HALF_TO_EVEN> {
-  template <typename T = Type>
-  static constexpr enable_if_floating_value<T> Round(const T val) {
-    return std::round(val * T(0.5)) * 2;
-  }
-
-  template <typename T = Type>
-  static enable_if_decimal_value<T, void> Round(T* val, const T& remainder,
-                                                const T& pow10, const int32_t scale) {
-    auto scaled = val->ReduceScaleBy(scale, /*round=*/false);
-    if (scaled.low_bits() % 2 != 0) {
-      scaled += remainder.Sign() >= 0 ? 1 : -1;
-    }
-    *val = scaled.IncreaseScaleBy(scale);
-  }
-};
-
-template <typename Type>
-struct RoundImpl<Type, RoundMode::HALF_TO_ODD> {
-  template <typename T = Type>
-  static constexpr enable_if_floating_value<T> Round(const T val) {
-    return std::floor(val * T(0.5)) + std::ceil(val * T(0.5));
-  }
-
-  template <typename T = Type>
-  static enable_if_decimal_value<T, void> Round(T* val, const T& remainder,
-                                                const T& pow10, const int32_t scale) {
-    auto scaled = val->ReduceScaleBy(scale, /*round=*/false);
-    if (scaled.low_bits() % 2 == 0) {
-      scaled += remainder.Sign() ? 1 : -1;
-    }
-    *val = scaled.IncreaseScaleBy(scale);
-  }
-};
-
-// Specializations of kernel state for round kernels
-template <typename OptionsType>
-struct RoundOptionsWrapper;
-
-template <>
-struct RoundOptionsWrapper<RoundOptions> : public OptionsWrapper<RoundOptions> {
-  using OptionsType = RoundOptions;
-  double pow10;
-
-  explicit RoundOptionsWrapper(OptionsType options) : OptionsWrapper(std::move(options)) {
-    // Only positive exponents for powers of 10 are used because combining
-    // multiply and division operations produced more stable rounding than
-    // using multiply-only.  Refer to NumPy's round implementation:
-    // https://github.com/numpy/numpy/blob/7b2f20b406d27364c812f7a81a9c901afbd3600c/numpy/core/src/multiarray/calculation.c#L589
-    pow10 = RoundUtil::Pow10(std::abs(options.ndigits));
-  }
-
-  static Result<std::unique_ptr<KernelState>> Init(KernelContext* ctx,
-                                                   const KernelInitArgs& args) {
-    if (auto options = static_cast<const OptionsType*>(args.options)) {
-      return std::make_unique<RoundOptionsWrapper>(*options);
-    }
-    return Status::Invalid(
-        "Attempted to initialize KernelState from null FunctionOptions");
-  }
-};
-
-template <>
-struct RoundOptionsWrapper<RoundToMultipleOptions>
-    : public OptionsWrapper<RoundToMultipleOptions> {
-  using OptionsType = RoundToMultipleOptions;
-  using OptionsWrapper::OptionsWrapper;
-
-  static Result<std::unique_ptr<KernelState>> Init(KernelContext* ctx,
-                                                   const KernelInitArgs& args) {
-    auto options = static_cast<const OptionsType*>(args.options);
-    if (!options) {
-      return Status::Invalid(
-          "Attempted to initialize KernelState from null FunctionOptions");
-    }
-
-    const auto& multiple = options->multiple;
-    if (!multiple || !multiple->is_valid) {
-      return Status::Invalid("Rounding multiple must be non-null and valid");
-    }
-
-    if (!IsPositive(*multiple)) {
-      return Status::Invalid("Rounding multiple must be positive");
-    }
-
-    // Ensure the rounding multiple option matches the kernel's output type.
-    // The output type is not available here so we use the following rule:
-    // If `multiple` is neither a floating-point nor a decimal type, then
-    // cast to float64, else cast to the kernel's input type.
-    std::shared_ptr<DataType> to_type =
-        (!is_floating(multiple->type->id()) && !is_decimal(multiple->type->id()))
-            ? float64()
-            : args.inputs[0].GetSharedPtr();
-    if (!multiple->type->Equals(to_type)) {
-      ARROW_ASSIGN_OR_RAISE(
-          auto casted_multiple,
-          Cast(Datum(multiple), to_type, CastOptions::Safe(), ctx->exec_context()));
-
-      // Create a new option object if the rounding multiple was casted.
-      auto new_options = OptionsType(casted_multiple.scalar(), options->round_mode);
-      return std::make_unique<RoundOptionsWrapper>(new_options);
-    }
-
-    return std::make_unique<RoundOptionsWrapper>(*options);
-  }
-};
-
-template <typename ArrowType, RoundMode RndMode, typename Enable = void>
-struct Round {
-  using CType = typename TypeTraits<ArrowType>::CType;
-  using State = RoundOptionsWrapper<RoundOptions>;
-
-  CType pow10;
-  int64_t ndigits;
-
-  explicit Round(const State& state, const DataType& out_ty)
-      : pow10(static_cast<CType>(state.pow10)), ndigits(state.options.ndigits) {}
-
-  template <typename T = ArrowType, typename CType = typename TypeTraits<T>::CType>
-  enable_if_floating_value<CType> Call(KernelContext* ctx, CType arg, Status* st) const {
-    // Do not process Inf or NaN because they will trigger the overflow error at end of
-    // function.
-    if (!std::isfinite(arg)) {
-      return arg;
-    }
-    auto round_val = ndigits >= 0 ? (arg * pow10) : (arg / pow10);
-    auto frac = round_val - std::floor(round_val);
-    if (frac != T(0)) {
-      // Use std::round() if in tie-breaking mode and scaled value is not 0.5.
-      if ((RndMode >= RoundMode::HALF_DOWN) && (frac != T(0.5))) {
-        round_val = std::round(round_val);
-      } else {
-        round_val = RoundImpl<CType, RndMode>::Round(round_val);
-      }
-      // Equality check is ommitted so that the common case of 10^0 (integer rounding)
-      // uses multiply-only
-      round_val = ndigits > 0 ? (round_val / pow10) : (round_val * pow10);
-      if (!std::isfinite(round_val)) {
-        *st = Status::Invalid("overflow occurred during rounding");
-        return arg;
-      }
-    } else {
-      // If scaled value is an integer, then no rounding is needed.
-      round_val = arg;
-    }
-    return round_val;
-  }
-};
-
-template <typename ArrowType, RoundMode kRoundMode>
-struct Round<ArrowType, kRoundMode, enable_if_decimal<ArrowType>> {
-  using CType = typename TypeTraits<ArrowType>::CType;
-  using State = RoundOptionsWrapper<RoundOptions>;
-
-  const ArrowType& ty;
-  int64_t ndigits;
-  int32_t pow;
-  // pow10 is "1" for the given decimal scale. Similarly half_pow10 is "0.5".
-  CType pow10, half_pow10, neg_half_pow10;
-
-  explicit Round(const State& state, const DataType& out_ty)
-      : Round(state.options.ndigits, out_ty) {}
-
-  explicit Round(int64_t ndigits, const DataType& out_ty)
-      : ty(checked_cast<const ArrowType&>(out_ty)),
-        ndigits(ndigits),
-        pow(static_cast<int32_t>(ty.scale() - ndigits)) {
-    if (pow >= ty.precision() || pow < 0) {
-      pow10 = half_pow10 = neg_half_pow10 = 0;
-    } else {
-      pow10 = CType::GetScaleMultiplier(pow);
-      half_pow10 = CType::GetHalfScaleMultiplier(pow);
-      neg_half_pow10 = -half_pow10;
-    }
-  }
-
-  template <typename T = ArrowType, typename CType = typename TypeTraits<T>::CType>
-  enable_if_decimal_value<CType> Call(KernelContext* ctx, CType arg, Status* st) const {
-    if (pow >= ty.precision()) {
-      *st = Status::Invalid("Rounding to ", ndigits,
-                            " digits will not fit in precision of ", ty);
-      return 0;
-    } else if (pow < 0) {
-      // no-op, copy output to input
-      return arg;
-    }
-
-    std::pair<CType, CType> pair;
-    *st = arg.Divide(pow10).Value(&pair);
-    if (!st->ok()) return arg;
-    // The remainder is effectively the scaled fractional part after division.
-    const auto& remainder = pair.second;
-    if (remainder == 0) return arg;
-    if (kRoundMode >= RoundMode::HALF_DOWN) {
-      if (remainder == half_pow10 || remainder == neg_half_pow10) {
-        // On the halfway point, use tiebreaker
-        RoundImpl<CType, kRoundMode>::Round(&arg, remainder, pow10, pow);
-      } else if (remainder.Sign() >= 0) {
-        // Positive, round up/down
-        arg -= remainder;
-        if (remainder > half_pow10) {
-          arg += pow10;
-        }
-      } else {
-        // Negative, round up/down
-        arg -= remainder;
-        if (remainder < neg_half_pow10) {
-          arg -= pow10;
-        }
-      }
-    } else {
-      RoundImpl<CType, kRoundMode>::Round(&arg, remainder, pow10, pow);
-    }
-    if (!arg.FitsInPrecision(ty.precision())) {
-      *st = Status::Invalid("Rounded value ", arg.ToString(ty.scale()),
-                            " does not fit in precision of ", ty);
-      return 0;
-    }
-    return arg;
-  }
-};
-
-template <typename DecimalType, RoundMode kMode, int32_t kDigits>
-Status FixedRoundDecimalExec(KernelContext* ctx, const ExecSpan& batch, ExecResult* out) {
-  using Op = Round<DecimalType, kMode>;
-  return ScalarUnaryNotNullStateful<DecimalType, DecimalType, Op>(
-             Op(kDigits, *out->type()))
-      .Exec(ctx, batch, out);
-}
-
-template <typename ArrowType, RoundMode kRoundMode, typename Enable = void>
-struct RoundToMultiple {
-  using CType = typename TypeTraits<ArrowType>::CType;
-  using State = RoundOptionsWrapper<RoundToMultipleOptions>;
-
-  CType multiple;
-
-  explicit RoundToMultiple(const State& state, const DataType& out_ty)
-      : multiple(UnboxScalar<ArrowType>::Unbox(*state.options.multiple)) {
-    const auto& options = state.options;
-    DCHECK(options.multiple);
-    DCHECK(options.multiple->is_valid);
-    DCHECK(is_floating(options.multiple->type->id()));
-  }
-
-  template <typename T = ArrowType, typename CType = typename TypeTraits<T>::CType>
-  enable_if_floating_value<CType> Call(KernelContext* ctx, CType arg, Status* st) const {
-    // Do not process Inf or NaN because they will trigger the overflow error at end of
-    // function.
-    if (!std::isfinite(arg)) {
-      return arg;
-    }
-    auto round_val = arg / multiple;
-    auto frac = round_val - std::floor(round_val);
-    if (frac != T(0)) {
-      // Use std::round() if in tie-breaking mode and scaled value is not 0.5.
-      if ((kRoundMode >= RoundMode::HALF_DOWN) && (frac != T(0.5))) {
-        round_val = std::round(round_val);
-      } else {
-        round_val = RoundImpl<CType, kRoundMode>::Round(round_val);
-      }
-      round_val *= multiple;
-      if (!std::isfinite(round_val)) {
-        *st = Status::Invalid("overflow occurred during rounding");
-        return arg;
-      }
-    } else {
-      // If scaled value is an integer, then no rounding is needed.
-      round_val = arg;
-    }
-    return round_val;
-  }
-};
-
-template <typename ArrowType, RoundMode kRoundMode>
-struct RoundToMultiple<ArrowType, kRoundMode, enable_if_decimal<ArrowType>> {
-  using CType = typename TypeTraits<ArrowType>::CType;
-  using State = RoundOptionsWrapper<RoundToMultipleOptions>;
-
-  const ArrowType& ty;
-  CType multiple, half_multiple, neg_half_multiple;
-  bool has_halfway_point;
-
-  explicit RoundToMultiple(const State& state, const DataType& out_ty)
-      : ty(checked_cast<const ArrowType&>(out_ty)),
-        multiple(UnboxScalar<ArrowType>::Unbox(*state.options.multiple)),
-        half_multiple(multiple / 2),
-        neg_half_multiple(-half_multiple),
-        has_halfway_point(multiple.low_bits() % 2 == 0) {
-    const auto& options = state.options;
-    DCHECK(options.multiple);
-    DCHECK(options.multiple->is_valid);
-    DCHECK(options.multiple->type->Equals(out_ty));
-  }
-
-  template <typename T = ArrowType, typename CType = typename TypeTraits<T>::CType>
-  enable_if_decimal_value<CType> Call(KernelContext* ctx, CType arg, Status* st) const {
-    std::pair<CType, CType> pair;
-    *st = arg.Divide(multiple).Value(&pair);
-    if (!st->ok()) return arg;
-    const auto& remainder = pair.second;
-    if (remainder == 0) return arg;
-    if (kRoundMode >= RoundMode::HALF_DOWN) {
-      if (has_halfway_point &&
-          (remainder == half_multiple || remainder == neg_half_multiple)) {
-        // On the halfway point, use tiebreaker
-        // Manually implement rounding since we're not actually rounding a
-        // decimal value, but rather manipulating the multiple
-        switch (kRoundMode) {
-          case RoundMode::HALF_DOWN:
-            if (remainder.Sign() < 0) pair.first -= 1;
-            break;
-          case RoundMode::HALF_UP:
-            if (remainder.Sign() >= 0) pair.first += 1;
-            break;
-          case RoundMode::HALF_TOWARDS_ZERO:
-            // Do nothing
-            break;
-          case RoundMode::HALF_TOWARDS_INFINITY:
-            pair.first += remainder.Sign() >= 0 ? 1 : -1;
-            break;
-          case RoundMode::HALF_TO_EVEN:
-            if (pair.first.low_bits() % 2 != 0) {
-              pair.first += remainder.Sign() >= 0 ? 1 : -1;
-            }
-            break;
-          case RoundMode::HALF_TO_ODD:
-            if (pair.first.low_bits() % 2 == 0) {
-              pair.first += remainder.Sign() >= 0 ? 1 : -1;
-            }
-            break;
-          default:
-            DCHECK(false);
-        }
-      } else if (remainder.Sign() >= 0) {
-        // Positive, round up/down
-        if (remainder > half_multiple) {
-          pair.first += 1;
-        }
-      } else {
-        // Negative, round up/down
-        if (remainder < neg_half_multiple) {
-          pair.first -= 1;
-        }
-      }
-    } else {
-      // Manually implement rounding since we're not actually rounding a
-      // decimal value, but rather manipulating the multiple
-      switch (kRoundMode) {
-        case RoundMode::DOWN:
-          if (remainder.Sign() < 0) pair.first -= 1;
-          break;
-        case RoundMode::UP:
-          if (remainder.Sign() >= 0) pair.first += 1;
-          break;
-        case RoundMode::TOWARDS_ZERO:
-          // Do nothing
-          break;
-        case RoundMode::TOWARDS_INFINITY:
-          pair.first += remainder.Sign() >= 0 ? 1 : -1;
-          break;
-        default:
-          DCHECK(false);
-      }
-    }
-    CType round_val = pair.first * multiple;
-    if (!round_val.FitsInPrecision(ty.precision())) {
-      *st = Status::Invalid("Rounded value ", round_val.ToString(ty.scale()),
-                            " does not fit in precision of ", ty);
-      return 0;
-    }
-    return round_val;
-  }
-};
-
-struct Floor {
-  template <typename T, typename Arg>
-  static constexpr enable_if_floating_value<Arg, T> Call(KernelContext*, Arg arg,
-                                                         Status*) {
-    static_assert(std::is_same<T, Arg>::value, "");
-    return RoundImpl<T, RoundMode::DOWN>::Round(arg);
-  }
-};
-
-struct Ceil {
-  template <typename T, typename Arg>
-  static constexpr enable_if_floating_value<Arg, T> Call(KernelContext*, Arg arg,
-                                                         Status*) {
-    static_assert(std::is_same<T, Arg>::value, "");
-    return RoundImpl<T, RoundMode::UP>::Round(arg);
-  }
-};
-
-struct Trunc {
-  template <typename T, typename Arg>
-  static constexpr enable_if_floating_value<Arg, T> Call(KernelContext*, Arg arg,
-                                                         Status*) {
-    static_assert(std::is_same<T, Arg>::value, "");
-    return RoundImpl<T, RoundMode::TOWARDS_ZERO>::Round(arg);
   }
 };
 
@@ -1416,77 +840,6 @@ std::shared_ptr<ScalarFunction> MakeUnaryArithmeticFunctionNotNull(std::string n
   for (const auto& ty : NumericTypes()) {
     auto exec = ArithmeticExecFromOp<ScalarUnaryNotNull, Op>(ty);
     DCHECK_OK(func->AddKernel({ty}, ty, exec));
-  }
-  AddNullExec(func.get());
-  return func;
-}
-
-#define ROUND_CASE(MODE)                                                       \
-  case RoundMode::MODE: {                                                      \
-    using Op = OpImpl<Type, RoundMode::MODE>;                                  \
-    return ScalarUnaryNotNullStateful<Type, Type, Op>(Op(state, *out->type())) \
-        .Exec(ctx, batch, out);                                                \
-  }
-
-// Exec the round kernel for the given types
-template <typename Type, typename OptionsType,
-          template <typename, RoundMode, typename...> class OpImpl>
-struct RoundKernel {
-  static Status Exec(KernelContext* ctx, const ExecSpan& batch, ExecResult* out) {
-    using State = RoundOptionsWrapper<OptionsType>;
-    const auto& state = static_cast<const State&>(*ctx->state());
-    switch (state.options.round_mode) {
-      ROUND_CASE(DOWN)
-      ROUND_CASE(UP)
-      ROUND_CASE(TOWARDS_ZERO)
-      ROUND_CASE(TOWARDS_INFINITY)
-      ROUND_CASE(HALF_DOWN)
-      ROUND_CASE(HALF_UP)
-      ROUND_CASE(HALF_TOWARDS_ZERO)
-      ROUND_CASE(HALF_TOWARDS_INFINITY)
-      ROUND_CASE(HALF_TO_EVEN)
-      ROUND_CASE(HALF_TO_ODD)
-    }
-    DCHECK(false);
-    return Status::NotImplemented(
-        "Internal implementation error: round mode not implemented: ",
-        state.options.ToString());
-  }
-};
-#undef ROUND_CASE
-
-// Like MakeUnaryArithmeticFunction, but for unary rounding functions that control
-// kernel dispatch based on RoundMode, only on non-null output.
-template <template <typename, RoundMode, typename...> class Op, typename OptionsType>
-std::shared_ptr<ScalarFunction> MakeUnaryRoundFunction(std::string name,
-                                                       FunctionDoc doc) {
-  using State = RoundOptionsWrapper<OptionsType>;
-  static const OptionsType kDefaultOptions = OptionsType::Defaults();
-  auto func = std::make_shared<ArithmeticIntegerToFloatingPointFunction>(
-      name, Arity::Unary(), std::move(doc), &kDefaultOptions);
-  for (const auto& ty : {float32(), float64(), decimal128(1, 0), decimal256(1, 0)}) {
-    auto type_id = ty->id();
-    ArrayKernelExec exec = nullptr;
-    switch (type_id) {
-      case Type::FLOAT:
-        exec = RoundKernel<FloatType, OptionsType, Op>::Exec;
-        break;
-      case Type::DOUBLE:
-        exec = RoundKernel<DoubleType, OptionsType, Op>::Exec;
-        break;
-      case Type::DECIMAL128:
-        exec = RoundKernel<Decimal128Type, OptionsType, Op>::Exec;
-        break;
-      case Type::DECIMAL256:
-        exec = RoundKernel<Decimal256Type, OptionsType, Op>::Exec;
-        break;
-      default:
-        DCHECK(false);
-        break;
-    }
-    DCHECK_OK(func->AddKernel(
-        {InputType(type_id)},
-        is_decimal(type_id) ? OutputType(FirstType) : OutputType(ty), exec, State::Init));
   }
   AddNullExec(func.get());
   return func;
@@ -1925,36 +1278,6 @@ const FunctionDoc logb_checked_doc{
      "-inf or NaN."),
     {"x", "b"}};
 
-const FunctionDoc floor_doc{
-    "Round down to the nearest integer",
-    ("Compute the largest integer value not greater in magnitude than `x`."),
-    {"x"}};
-
-const FunctionDoc ceil_doc{
-    "Round up to the nearest integer",
-    ("Compute the smallest integer value not less in magnitude than `x`."),
-    {"x"}};
-
-const FunctionDoc trunc_doc{
-    "Compute the integral part",
-    ("Compute the nearest integer not greater in magnitude than `x`."),
-    {"x"}};
-
-const FunctionDoc round_doc{
-    "Round to a given precision",
-    ("Options are used to control the number of digits and rounding mode.\n"
-     "Default behavior is to round to the nearest integer and\n"
-     "use half-to-even rule to break ties."),
-    {"x"},
-    "RoundOptions"};
-
-const FunctionDoc round_to_multiple_doc{
-    "Round to a given multiple",
-    ("Options are used to control the rounding multiple and rounding mode.\n"
-     "Default behavior is to round to the nearest integer and\n"
-     "use half-to-even rule to break ties."),
-    {"x"},
-    "RoundToMultipleOptions"};
 }  // namespace
 
 void RegisterScalarArithmetic(FunctionRegistry* registry) {
@@ -2365,52 +1688,6 @@ void RegisterScalarArithmetic(FunctionRegistry* registry) {
   auto logb_checked = MakeArithmeticFunctionFloatingPointNotNull<LogbChecked>(
       "logb_checked", logb_checked_doc);
   DCHECK_OK(registry->AddFunction(std::move(logb_checked)));
-
-  // ----------------------------------------------------------------------
-  // Rounding functions
-  auto floor =
-      MakeUnaryArithmeticFunctionFloatingPoint<Floor,
-                                               ArithmeticIntegerToFloatingPointFunction>(
-          "floor", floor_doc);
-  DCHECK_OK(floor->AddKernel(
-      {InputType(Type::DECIMAL128)}, OutputType(FirstType),
-      FixedRoundDecimalExec<Decimal128Type, RoundMode::DOWN, /*ndigits=*/0>));
-  DCHECK_OK(floor->AddKernel(
-      {InputType(Type::DECIMAL256)}, OutputType(FirstType),
-      FixedRoundDecimalExec<Decimal256Type, RoundMode::DOWN, /*ndigits=*/0>));
-  DCHECK_OK(registry->AddFunction(std::move(floor)));
-
-  auto ceil =
-      MakeUnaryArithmeticFunctionFloatingPoint<Ceil,
-                                               ArithmeticIntegerToFloatingPointFunction>(
-          "ceil", ceil_doc);
-  DCHECK_OK(ceil->AddKernel(
-      {InputType(Type::DECIMAL128)}, OutputType(FirstType),
-      FixedRoundDecimalExec<Decimal128Type, RoundMode::UP, /*ndigits=*/0>));
-  DCHECK_OK(ceil->AddKernel(
-      {InputType(Type::DECIMAL256)}, OutputType(FirstType),
-      FixedRoundDecimalExec<Decimal256Type, RoundMode::UP, /*ndigits=*/0>));
-  DCHECK_OK(registry->AddFunction(std::move(ceil)));
-
-  auto trunc =
-      MakeUnaryArithmeticFunctionFloatingPoint<Trunc,
-                                               ArithmeticIntegerToFloatingPointFunction>(
-          "trunc", trunc_doc);
-  DCHECK_OK(trunc->AddKernel(
-      {InputType(Type::DECIMAL128)}, OutputType(FirstType),
-      FixedRoundDecimalExec<Decimal128Type, RoundMode::TOWARDS_ZERO, /*ndigits=*/0>));
-  DCHECK_OK(trunc->AddKernel(
-      {InputType(Type::DECIMAL256)}, OutputType(FirstType),
-      FixedRoundDecimalExec<Decimal256Type, RoundMode::TOWARDS_ZERO, /*ndigits=*/0>));
-  DCHECK_OK(registry->AddFunction(std::move(trunc)));
-
-  auto round = MakeUnaryRoundFunction<Round, RoundOptions>("round", round_doc);
-  DCHECK_OK(registry->AddFunction(std::move(round)));
-
-  auto round_to_multiple =
-      MakeUnaryRoundFunction<RoundToMultiple, RoundToMultipleOptions>(
-          "round_to_multiple", round_to_multiple_doc);
-  DCHECK_OK(registry->AddFunction(std::move(round_to_multiple)));
 }
 
 }  // namespace internal
