+++ cpp/src/arrow/compute/kernels/scalar_arithmetic_test.cc
@@ -1238,8 +1238,7 @@ TEST(TestUnaryArithmetic, DispatchBest) {
   }
 
   // Float types
-  for (std::string name :
-       {"atan", "sign", "floor", "ceil", "trunc", "round", "round_to_multiple", "exp"}) {
+  for (std::string name : {"atan", "sign", "exp"}) {
     for (const auto& ty : {float32(), float64()}) {
       CheckDispatchBest(name, {ty}, {ty});
       CheckDispatchBest(name, {dictionary(int8(), ty)}, {ty});
@@ -1260,8 +1259,7 @@ TEST(TestUnaryArithmetic, DispatchBest) {
   }
 
   // Integer -> Float64
-  for (std::string name :
-       {"atan", "floor", "ceil", "trunc", "round", "round_to_multiple"}) {
+  for (std::string name : {"atan"}) {
     for (const auto& ty :
          {int8(), int16(), int32(), int64(), uint8(), uint16(), uint32(), uint64()}) {
       CheckDispatchBest(name, {ty}, {float64()});
@@ -1691,504 +1689,6 @@ TEST_F(TestUnaryArithmeticDecimal, Negate) {
   }
 }
 
-// Check two modes exhaustively, give all modes a simple test
-TEST_F(TestUnaryArithmeticDecimal, Round) {
-  const auto func = "round";
-  RoundOptions options(2, RoundMode::DOWN);
-  for (const auto& ty : {decimal128(4, 3), decimal256(4, 3)}) {
-    auto values = ArrayFromJSON(
-        ty,
-        R"(["1.010", "1.012", "1.015", "1.019", "-1.010", "-1.012", "-1.015", "-1.019", null])");
-    options.round_mode = RoundMode::DOWN;
-    CheckScalar(
-        func, {values},
-        ArrayFromJSON(
-            ty,
-            R"(["1.010", "1.010", "1.010", "1.010", "-1.010", "-1.020", "-1.020", "-1.020", null])"),
-        &options);
-    options.round_mode = RoundMode::UP;
-    CheckScalar(
-        func, {values},
-        ArrayFromJSON(
-            ty,
-            R"(["1.010", "1.020", "1.020", "1.020", "-1.010", "-1.010", "-1.010", "-1.010", null])"),
-        &options);
-    options.round_mode = RoundMode::TOWARDS_ZERO;
-    CheckScalar(
-        func, {values},
-        ArrayFromJSON(
-            ty,
-            R"(["1.010", "1.010", "1.010", "1.010", "-1.010", "-1.010", "-1.010", "-1.010", null])"),
-        &options);
-    options.round_mode = RoundMode::TOWARDS_INFINITY;
-    CheckScalar(
-        func, {values},
-        ArrayFromJSON(
-            ty,
-            R"(["1.010", "1.020", "1.020", "1.020", "-1.010", "-1.020", "-1.020", "-1.020", null])"),
-        &options);
-    options.round_mode = RoundMode::HALF_DOWN;
-    CheckScalar(
-        func, {values},
-        ArrayFromJSON(
-            ty,
-            R"(["1.010", "1.010", "1.010", "1.020", "-1.010", "-1.010", "-1.020", "-1.020", null])"),
-        &options);
-    options.round_mode = RoundMode::HALF_UP;
-    CheckScalar(
-        func, {values},
-        ArrayFromJSON(
-            ty,
-            R"(["1.010", "1.010", "1.020", "1.020", "-1.010", "-1.010", "-1.010", "-1.020", null])"),
-        &options);
-    options.round_mode = RoundMode::HALF_TOWARDS_ZERO;
-    CheckScalar(
-        func, {values},
-        ArrayFromJSON(
-            ty,
-            R"(["1.010", "1.010", "1.010", "1.020", "-1.010", "-1.010", "-1.010", "-1.020", null])"),
-        &options);
-    options.round_mode = RoundMode::HALF_TOWARDS_INFINITY;
-    CheckScalar(
-        func, {values},
-        ArrayFromJSON(
-            ty,
-            R"(["1.010", "1.010", "1.020", "1.020", "-1.010", "-1.010", "-1.020", "-1.020", null])"),
-        &options);
-    options.round_mode = RoundMode::HALF_TO_EVEN;
-    CheckScalar(
-        func, {values},
-        ArrayFromJSON(
-            ty,
-            R"(["1.010", "1.010", "1.020", "1.020", "-1.010", "-1.010", "-1.020", "-1.020", null])"),
-        &options);
-    options.round_mode = RoundMode::HALF_TO_ODD;
-    CheckScalar(
-        func, {values},
-        ArrayFromJSON(
-            ty,
-            R"(["1.010", "1.010", "1.010", "1.020", "-1.010", "-1.010", "-1.010", "-1.020", null])"),
-        &options);
-  }
-}
-
-TEST_F(TestUnaryArithmeticDecimal, RoundTowardsInfinity) {
-  const auto func = "round";
-  RoundOptions options(0, RoundMode::TOWARDS_INFINITY);
-  for (const auto& ty : {decimal128(4, 2), decimal256(4, 2)}) {
-    auto values = ArrayFromJSON(
-        ty, R"(["1.00", "1.99", "1.01", "-42.00", "-42.99", "-42.15", null])");
-    CheckScalar(func, {ArrayFromJSON(ty, R"([])")}, ArrayFromJSON(ty, R"([])"), &options);
-    options.ndigits = 0;
-    CheckScalar(
-        func, {values},
-        ArrayFromJSON(ty,
-                      R"(["1.00", "2.00", "2.00", "-42.00", "-43.00", "-43.00", null])"),
-        &options);
-    options.ndigits = 1;
-    CheckScalar(
-        func, {values},
-        ArrayFromJSON(ty,
-                      R"(["1.00", "2.00", "1.10", "-42.00", "-43.00", "-42.20", null])"),
-        &options);
-    options.ndigits = 2;
-    CheckScalar(func, {values}, values, &options);
-    options.ndigits = 4;
-    CheckScalar(func, {values}, values, &options);
-    options.ndigits = 100;
-    CheckScalar(func, {values}, values, &options);
-    options.ndigits = -1;
-    CheckScalar(
-        func, {values},
-        ArrayFromJSON(
-            ty, R"(["10.00", "10.00", "10.00", "-50.00", "-50.00", "-50.00", null])"),
-        &options);
-    options.ndigits = -2;
-    CheckRaises(func, {values}, "Rounding to -2 digits will not fit in precision",
-                &options);
-    options.ndigits = -1;
-    CheckRaises(func, {ArrayFromJSON(ty, R"(["99.99"])")},
-                "Rounded value 100.00 does not fit in precision", &options);
-  }
-  for (const auto& ty : {decimal128(2, -2), decimal256(2, -2)}) {
-    auto values = DecimalArrayFromJSON(
-        ty, R"(["10E2", "12E2", "18E2", "-10E2", "-12E2", "-18E2", null])");
-    options.ndigits = 0;
-    CheckScalar(func, {values}, values, &options);
-    options.ndigits = 2;
-    CheckScalar(func, {values}, values, &options);
-    options.ndigits = 100;
-    CheckScalar(func, {values}, values, &options);
-    options.ndigits = -1;
-    CheckScalar(func, {values}, values, &options);
-    options.ndigits = -2;
-    CheckScalar(func, {values}, values, &options);
-    options.ndigits = -3;
-    CheckScalar(func, {values},
-                DecimalArrayFromJSON(
-                    ty, R"(["10E2", "20E2", "20E2", "-10E2", "-20E2", "-20E2", null])"),
-                &options);
-    options.ndigits = -4;
-    CheckRaises(func, {values}, "Rounding to -4 digits will not fit in precision",
-                &options);
-  }
-}
-
-TEST_F(TestUnaryArithmeticDecimal, RoundHalfToEven) {
-  const auto func = "round";
-  RoundOptions options(0, RoundMode::HALF_TO_EVEN);
-  for (const auto& ty : {decimal128(4, 2), decimal256(4, 2)}) {
-    auto values = ArrayFromJSON(
-        ty,
-        R"(["1.00", "5.99", "1.01", "-42.00", "-42.99", "-42.15", "1.50", "2.50", "-5.50", "-2.55", null])");
-    CheckScalar(func, {ArrayFromJSON(ty, R"([])")}, ArrayFromJSON(ty, R"([])"), &options);
-    options.ndigits = 0;
-    CheckScalar(
-        func, {values},
-        ArrayFromJSON(
-            ty,
-            R"(["1.00", "6.00", "1.00", "-42.00", "-43.00", "-42.00", "2.00", "2.00", "-6.00", "-3.00", null])"),
-        &options);
-    options.ndigits = 1;
-    CheckScalar(
-        func, {values},
-        ArrayFromJSON(
-            ty,
-            R"(["1.00", "6.00", "1.00", "-42.00", "-43.00", "-42.20", "1.50", "2.50", "-5.50", "-2.60", null])"),
-        &options);
-    options.ndigits = 2;
-    CheckScalar(func, {values}, values, &options);
-    options.ndigits = 4;
-    CheckScalar(func, {values}, values, &options);
-    options.ndigits = 100;
-    CheckScalar(func, {values}, values, &options);
-    options.ndigits = -1;
-    CheckScalar(
-        func, {values},
-        ArrayFromJSON(
-            ty,
-            R"(["0.00", "10.00", "0.00", "-40.00", "-40.00", "-40.00", "0.00", "0.00", "-10.00", "0.00", null])"),
-        &options);
-    options.ndigits = -2;
-    CheckRaises(func, {values}, "Rounding to -2 digits will not fit in precision",
-                &options);
-    options.ndigits = -1;
-    CheckRaises(func, {ArrayFromJSON(ty, R"(["99.99"])")},
-                "Rounded value 100.00 does not fit in precision", &options);
-  }
-  for (const auto& ty : {decimal128(2, -2), decimal256(2, -2)}) {
-    auto values = DecimalArrayFromJSON(
-        ty,
-        R"(["5E2", "10E2", "12E2", "15E2", "18E2", "-10E2", "-12E2", "-15E2", "-18E2", null])");
-    options.ndigits = 0;
-    CheckScalar(func, {values}, values, &options);
-    options.ndigits = 2;
-    CheckScalar(func, {values}, values, &options);
-    options.ndigits = 100;
-    CheckScalar(func, {values}, values, &options);
-    options.ndigits = -1;
-    CheckScalar(func, {values}, values, &options);
-    options.ndigits = -2;
-    CheckScalar(func, {values}, values, &options);
-    options.ndigits = -3;
-    CheckScalar(
-        func, {values},
-        DecimalArrayFromJSON(
-            ty,
-            R"(["0", "10E2", "10E2", "20E2", "20E2", "-10E2", "-10E2", "-20E2", "-20E2", null])"),
-        &options);
-    options.ndigits = -4;
-    CheckRaises(func, {values}, "Rounding to -4 digits will not fit in precision",
-                &options);
-  }
-}
-
-TEST_F(TestUnaryArithmeticDecimal, RoundCeil) {
-  const auto func = "ceil";
-  for (const auto& ty : PositiveScaleTypes()) {
-    CheckScalar(func, {ArrayFromJSON(ty, R"([])")}, ArrayFromJSON(ty, R"([])"));
-    CheckScalar(
-        func,
-        {ArrayFromJSON(
-            ty, R"(["1.00", "1.99", "1.01", "-42.00", "-42.99", "-42.15", null])")},
-        ArrayFromJSON(ty,
-                      R"(["1.00", "2.00", "2.00", "-42.00", "-42.00", "-42.00", null])"));
-  }
-  for (const auto& ty : {decimal128(4, 2), decimal256(4, 2)}) {
-    CheckRaises(func, {ScalarFromJSON(ty, R"("99.99")")},
-                "Rounded value 100.00 does not fit in precision of decimal");
-    CheckScalar(func, {ScalarFromJSON(ty, R"("-99.99")")},
-                ScalarFromJSON(ty, R"("-99.00")"));
-  }
-  for (const auto& ty : NegativeScaleTypes()) {
-    CheckScalar(func, {ArrayFromJSON(ty, R"([])")}, ArrayFromJSON(ty, R"([])"));
-    CheckScalar(func, {DecimalArrayFromJSON(ty, R"(["12E2", "-42E2", null])")},
-                DecimalArrayFromJSON(ty, R"(["12E2", "-42E2", null])"));
-  }
-}
-
-TEST_F(TestUnaryArithmeticDecimal, RoundFloor) {
-  const auto func = "floor";
-  for (const auto& ty : PositiveScaleTypes()) {
-    CheckScalar(func, {ArrayFromJSON(ty, R"([])")}, ArrayFromJSON(ty, R"([])"));
-    CheckScalar(
-        func,
-        {ArrayFromJSON(
-            ty, R"(["1.00", "1.99", "1.01", "-42.00", "-42.99", "-42.15", null])")},
-        ArrayFromJSON(ty,
-                      R"(["1.00", "1.00", "1.00", "-42.00", "-43.00", "-43.00", null])"));
-  }
-  for (const auto& ty : {decimal128(4, 2), decimal256(4, 2)}) {
-    CheckScalar(func, {ScalarFromJSON(ty, R"("99.99")")},
-                ScalarFromJSON(ty, R"("99.00")"));
-    CheckRaises(func, {ScalarFromJSON(ty, R"("-99.99")")},
-                "Rounded value -100.00 does not fit in precision of decimal");
-  }
-  for (const auto& ty : NegativeScaleTypes()) {
-    CheckScalar(func, {ArrayFromJSON(ty, R"([])")}, ArrayFromJSON(ty, R"([])"));
-    CheckScalar(func, {DecimalArrayFromJSON(ty, R"(["12E2", "-42E2", null])")},
-                DecimalArrayFromJSON(ty, R"(["12E2", "-42E2", null])"));
-  }
-}
-
-TEST_F(TestUnaryArithmeticDecimal, RoundTrunc) {
-  const auto func = "trunc";
-  for (const auto& ty : PositiveScaleTypes()) {
-    CheckScalar(func, {ArrayFromJSON(ty, R"([])")}, ArrayFromJSON(ty, R"([])"));
-    CheckScalar(
-        func,
-        {ArrayFromJSON(
-            ty, R"(["1.00", "1.99", "1.01", "-42.00", "-42.99", "-42.15", null])")},
-        ArrayFromJSON(ty,
-                      R"(["1.00", "1.00", "1.00", "-42.00", "-42.00", "-42.00", null])"));
-  }
-  for (const auto& ty : {decimal128(4, 2), decimal256(4, 2)}) {
-    CheckScalar(func, {ScalarFromJSON(ty, R"("99.99")")},
-                ScalarFromJSON(ty, R"("99.00")"));
-    CheckScalar(func, {ScalarFromJSON(ty, R"("-99.99")")},
-                ScalarFromJSON(ty, R"("-99.00")"));
-  }
-  for (const auto& ty : NegativeScaleTypes()) {
-    CheckScalar(func, {ArrayFromJSON(ty, R"([])")}, ArrayFromJSON(ty, R"([])"));
-    CheckScalar(func, {DecimalArrayFromJSON(ty, R"(["12E2", "-42E2", null])")},
-                DecimalArrayFromJSON(ty, R"(["12E2", "-42E2", null])"));
-  }
-}
-
-TEST_F(TestUnaryArithmeticDecimal, RoundToMultiple) {
-  const auto func = "round_to_multiple";
-  RoundToMultipleOptions options(0, RoundMode::DOWN);
-  for (const auto& ty : {decimal128(4, 2), decimal256(4, 2)}) {
-    if (ty->id() == Type::DECIMAL128) {
-      options.multiple = std::make_shared<Decimal128Scalar>(Decimal128(200), ty);
-    } else {
-      options.multiple = std::make_shared<Decimal256Scalar>(Decimal256(200), ty);
-    }
-    auto values = ArrayFromJSON(
-        ty,
-        R"(["-3.50", "-3.00", "-2.50", "-2.00", "-1.50", "-1.00", "-0.50", "0.00",
-            "0.50", "1.00", "1.50", "2.00", "2.50", "3.00", "3.50", null])");
-    options.round_mode = RoundMode::DOWN;
-    CheckScalar(
-        func, {values},
-        ArrayFromJSON(
-            ty,
-            R"(["-4.00", "-4.00", "-4.00", "-2.00", "-2.00", "-2.00", "-2.00", "0.00",
-            "0.00", "0.00", "0.00", "2.00", "2.00", "2.00", "2.00", null])"),
-        &options);
-    options.round_mode = RoundMode::UP;
-    CheckScalar(
-        func, {values},
-        ArrayFromJSON(
-            ty,
-            R"(["-2.00", "-2.00", "-2.00", "-2.00", "-0.00", "-0.00", "-0.00", "0.00",
-            "2.00", "2.00", "2.00", "2.00", "4.00", "4.00", "4.00", null])"),
-        &options);
-    options.round_mode = RoundMode::TOWARDS_ZERO;
-    CheckScalar(
-        func, {values},
-        ArrayFromJSON(
-            ty,
-            R"(["-2.00", "-2.00", "-2.00", "-2.00", "-0.00", "-0.00", "-0.00", "0.00",
-            "0.00", "0.00", "0.00", "2.00", "2.00", "2.00", "2.00", null])"),
-        &options);
-    options.round_mode = RoundMode::TOWARDS_INFINITY;
-    CheckScalar(
-        func, {values},
-        ArrayFromJSON(
-            ty,
-            R"(["-4.00", "-4.00", "-4.00", "-2.00", "-2.00", "-2.00", "-2.00", "0.00",
-            "2.00", "2.00", "2.00", "2.00", "4.00", "4.00", "4.00", null])"),
-        &options);
-    options.round_mode = RoundMode::HALF_DOWN;
-    CheckScalar(
-        func, {values},
-        ArrayFromJSON(
-            ty,
-            R"(["-4.00", "-4.00", "-2.00", "-2.00", "-2.00", "-2.00", "-0.00", "0.00",
-            "0.00", "0.00", "2.00", "2.00", "2.00", "2.00", "4.00", null])"),
-        &options);
-    options.round_mode = RoundMode::HALF_UP;
-    CheckScalar(
-        func, {values},
-        ArrayFromJSON(
-            ty,
-            R"(["-4.00", "-2.00", "-2.00", "-2.00", "-2.00", "-0.00", "-0.00", "0.00",
-            "0.00", "2.00", "2.00", "2.00", "2.00", "4.00", "4.00", null])"),
-        &options);
-    options.round_mode = RoundMode::HALF_TOWARDS_ZERO;
-    CheckScalar(
-        func, {values},
-        ArrayFromJSON(
-            ty,
-            R"(["-4.00", "-2.00", "-2.00", "-2.00", "-2.00", "-0.00", "-0.00", "0.00",
-            "0.00", "0.00", "2.00", "2.00", "2.00", "2.00", "4.00", null])"),
-        &options);
-    options.round_mode = RoundMode::HALF_TOWARDS_INFINITY;
-    CheckScalar(
-        func, {values},
-        ArrayFromJSON(
-            ty,
-            R"(["-4.00", "-4.00", "-2.00", "-2.00", "-2.00", "-2.00", "-0.00", "0.00",
-            "0.00", "2.00", "2.00", "2.00", "2.00", "4.00", "4.00", null])"),
-        &options);
-    options.round_mode = RoundMode::HALF_TO_EVEN;
-    CheckScalar(
-        func, {values},
-        ArrayFromJSON(
-            ty,
-            R"(["-4.00", "-4.00", "-2.00", "-2.00", "-2.00", "-0.00", "-0.00", "0.00",
-            "0.00", "0.00", "2.00", "2.00", "2.00", "4.00", "4.00", null])"),
-        &options);
-    options.round_mode = RoundMode::HALF_TO_ODD;
-    CheckScalar(
-        func, {values},
-        ArrayFromJSON(
-            ty,
-            R"(["-4.00", "-2.00", "-2.00", "-2.00", "-2.00", "-2.00", "-0.00", "0.00",
-            "0.00", "2.00", "2.00", "2.00", "2.00", "2.00", "4.00", null])"),
-        &options);
-  }
-}
-
-TEST_F(TestUnaryArithmeticDecimal, RoundToMultipleTowardsInfinity) {
-  const auto func = "round_to_multiple";
-  RoundToMultipleOptions options(0, RoundMode::TOWARDS_INFINITY);
-  auto set_multiple = [&](const std::shared_ptr<DataType>& ty, int64_t value) {
-    if (ty->id() == Type::DECIMAL128) {
-      options.multiple = std::make_shared<Decimal128Scalar>(Decimal128(value), ty);
-    } else {
-      options.multiple = std::make_shared<Decimal256Scalar>(Decimal256(value), ty);
-    }
-  };
-  for (const auto& ty : {decimal128(4, 2), decimal256(4, 2)}) {
-    auto values = ArrayFromJSON(
-        ty, R"(["1.00", "1.99", "1.01", "-42.00", "-42.99", "-42.15", null])");
-    set_multiple(ty, 25);
-    CheckScalar(func, {ArrayFromJSON(ty, R"([])")}, ArrayFromJSON(ty, R"([])"), &options);
-    CheckScalar(
-        func, {values},
-        ArrayFromJSON(ty,
-                      R"(["1.00", "2.00", "1.25", "-42.00", "-43.00", "-42.25", null])"),
-        &options);
-    set_multiple(ty, 1);
-    CheckScalar(func, {values}, values, &options);
-    set_multiple(decimal128(2, 0), 2);
-    CheckScalar(
-        func, {values},
-        ArrayFromJSON(ty,
-                      R"(["2.00", "2.00", "2.00", "-42.00", "-44.00", "-44.00", null])"),
-        &options);
-    set_multiple(ty, 0);
-    CheckRaises(func, {values}, "Rounding multiple must be positive", &options);
-    options.multiple =
-        std::make_shared<Decimal128Scalar>(Decimal128(0), decimal128(4, 2));
-    CheckRaises(func, {values}, "Rounding multiple must be positive", &options);
-    set_multiple(ty, -10);
-    CheckRaises(func, {ArrayFromJSON(ty, R"(["99.99"])")},
-                "Rounding multiple must be positive", &options);
-    set_multiple(ty, 100);
-    CheckRaises(func, {ArrayFromJSON(ty, R"(["99.99"])")},
-                "Rounded value 100.00 does not fit in precision", &options);
-    options.multiple = std::make_shared<DoubleScalar>(1.0);
-    CheckRaises(func, {ArrayFromJSON(ty, R"(["99.99"])")},
-                "Rounded value 100.00 does not fit in precision", &options);
-    options.multiple = std::make_shared<Decimal128Scalar>(decimal128(3, 0));
-    CheckRaises(func, {ArrayFromJSON(ty, R"(["99.99"])")},
-                "Rounding multiple must be non-null and valid", &options);
-    options.multiple = nullptr;
-    CheckRaises(func, {ArrayFromJSON(ty, R"(["99.99"])")},
-                "Rounding multiple must be non-null and valid", &options);
-  }
-  for (const auto& ty : {decimal128(2, -2), decimal256(2, -2)}) {
-    auto values = DecimalArrayFromJSON(
-        ty, R"(["10E2", "12E2", "18E2", "-10E2", "-12E2", "-18E2", null])");
-    set_multiple(ty, 4);
-    CheckScalar(func, {values},
-                DecimalArrayFromJSON(
-                    ty, R"(["12E2", "12E2", "20E2", "-12E2", "-12E2", "-20E2", null])"),
-                &options);
-    set_multiple(ty, 1);
-    CheckScalar(func, {values}, values, &options);
-  }
-}
-
-TEST_F(TestUnaryArithmeticDecimal, RoundToMultipleHalfToOdd) {
-  const auto func = "round_to_multiple";
-  RoundToMultipleOptions options(0, RoundMode::HALF_TO_ODD);
-  auto set_multiple = [&](const std::shared_ptr<DataType>& ty, int64_t value) {
-    if (ty->id() == Type::DECIMAL128) {
-      options.multiple = std::make_shared<Decimal128Scalar>(Decimal128(value), ty);
-    } else {
-      options.multiple = std::make_shared<Decimal256Scalar>(Decimal256(value), ty);
-    }
-  };
-  for (const auto& ty : {decimal128(4, 2), decimal256(4, 2)}) {
-    auto values =
-        ArrayFromJSON(ty, R"(["-0.38", "-0.37", "-0.25", "-0.13", "-0.12", "0.00",
-                "0.12", "0.13", "0.25", "0.37", "0.38", null])");
-    // There is no exact halfway point, check what happens
-    set_multiple(ty, 25);
-    CheckScalar(func, {ArrayFromJSON(ty, R"([])")}, ArrayFromJSON(ty, R"([])"), &options);
-    CheckScalar(func, {values},
-                ArrayFromJSON(ty, R"(["-0.50", "-0.25", "-0.25", "-0.25", "-0.00", "0.00",
-                              "0.00", "0.25", "0.25", "0.25", "0.50", null])"),
-                &options);
-    set_multiple(ty, 1);
-    CheckScalar(func, {values}, values, &options);
-    set_multiple(ty, 24);
-    CheckScalar(func, {ArrayFromJSON(ty, R"([])")}, ArrayFromJSON(ty, R"([])"), &options);
-    CheckScalar(func, {values},
-                ArrayFromJSON(ty, R"(["-0.48", "-0.48", "-0.24", "-0.24", "-0.24", "0.00",
-                              "0.24", "0.24", "0.24", "0.48", "0.48", null])"),
-                &options);
-    set_multiple(decimal128(3, 1), 1);
-    CheckScalar(
-        func, {values},
-        ArrayFromJSON(
-            ty,
-            R"(["-0.40", "-0.40", "-0.30", "-0.10", "-0.10", "0.00", "0.10", "0.10",
-                      "0.30", "0.40", "0.40", null])"),
-        &options);
-  }
-  for (const auto& ty : {decimal128(2, -2), decimal256(2, -2)}) {
-    auto values = DecimalArrayFromJSON(
-        ty, R"(["10E2", "12E2", "18E2", "-10E2", "-12E2", "-18E2", null])");
-    set_multiple(ty, 4);
-    CheckScalar(func, {values},
-                DecimalArrayFromJSON(
-                    ty, R"(["12E2", "12E2", "20E2", "-12E2", "-12E2", "-20E2", null])"),
-                &options);
-    set_multiple(ty, 5);
-    CheckScalar(func, {values},
-                DecimalArrayFromJSON(
-                    ty, R"(["10E2", "10E2", "20E2", "-10E2", "-10E2", "-20E2", null])"),
-                &options);
-    set_multiple(ty, 1);
-    CheckScalar(func, {values}, values, &options);
-  }
-}
-
 TEST_F(TestUnaryArithmeticDecimal, Sign) {
   auto max128 = Decimal128::GetMaxValue(38);
   auto max256 = Decimal256::GetMaxValue(76);
@@ -2297,213 +1797,6 @@ TEST_F(TestUnaryArithmeticDecimal, TrigTan) {
       CheckDecimalToFloat(func, {DecimalArrayFromJSON(ty, R"(["12E2", "-42E2", null])")});
     }
   }
-}
-
-TYPED_TEST_SUITE(TestUnaryRoundIntegral, IntegralTypes);
-TYPED_TEST_SUITE(TestUnaryRoundSigned, SignedIntegerTypes);
-TYPED_TEST_SUITE(TestUnaryRoundUnsigned, UnsignedIntegerTypes);
-TYPED_TEST_SUITE(TestUnaryRoundFloating, FloatingTypes);
-
-const std::vector<RoundMode> kRoundModes{
-    RoundMode::DOWN,
-    RoundMode::UP,
-    RoundMode::TOWARDS_ZERO,
-    RoundMode::TOWARDS_INFINITY,
-    RoundMode::HALF_DOWN,
-    RoundMode::HALF_UP,
-    RoundMode::HALF_TOWARDS_ZERO,
-    RoundMode::HALF_TOWARDS_INFINITY,
-    RoundMode::HALF_TO_EVEN,
-    RoundMode::HALF_TO_ODD,
-};
-
-TYPED_TEST(TestUnaryRoundSigned, Round) {
-  // Test different rounding modes for integer rounding
-  std::string values("[0, 1, -13, -50, 115]");
-  this->SetRoundNdigits(0);
-  for (const auto& round_mode : kRoundModes) {
-    this->SetRoundMode(round_mode);
-    this->AssertUnaryOp(Round, values, ArrayFromJSON(float64(), values));
-  }
-
-  // Test different round N-digits for nearest rounding mode
-  std::vector<std::pair<int64_t, std::string>> ndigits_and_expected{{
-      {-2, "[0.0, 0.0, -0.0, -100, 100]"},
-      {-1, "[0.0, 0.0, -10, -50, 120]"},
-      {0, values},
-      {1, values},
-      {2, values},
-  }};
-  this->SetRoundMode(RoundMode::HALF_TOWARDS_INFINITY);
-  for (const auto& pair : ndigits_and_expected) {
-    this->SetRoundNdigits(pair.first);
-    this->AssertUnaryOp(Round, values, ArrayFromJSON(float64(), pair.second));
-  }
-}
-
-TYPED_TEST(TestUnaryRoundUnsigned, Round) {
-  // Test different rounding modes for integer rounding
-  std::string values("[0, 1, 13, 50, 115]");
-  this->SetRoundNdigits(0);
-  for (const auto& round_mode : kRoundModes) {
-    this->SetRoundMode(round_mode);
-    this->AssertUnaryOp(Round, values, ArrayFromJSON(float64(), values));
-  }
-
-  // Test different round N-digits for nearest rounding mode
-  std::vector<std::pair<int64_t, std::string>> ndigits_and_expected{{
-      {-2, "[0, 0, 0, 100, 100]"},
-      {-1, "[0, 0, 10, 50, 120]"},
-      {0, values},
-      {1, values},
-      {2, values},
-  }};
-  this->SetRoundMode(RoundMode::HALF_TOWARDS_INFINITY);
-  for (const auto& pair : ndigits_and_expected) {
-    this->SetRoundNdigits(pair.first);
-    this->AssertUnaryOp(Round, values, ArrayFromJSON(float64(), pair.second));
-  }
-}
-
-TYPED_TEST(TestUnaryRoundFloating, Round) {
-  this->SetNansEqual(true);
-
-  // Test different rounding modes
-  std::string values("[3.2, 3.5, 3.7, 4.5, -3.2, -3.5, -3.7]");
-  std::vector<std::pair<RoundMode, std::string>> rmode_and_expected{{
-      {RoundMode::DOWN, "[3, 3, 3, 4, -4, -4, -4]"},
-      {RoundMode::UP, "[4, 4, 4, 5, -3, -3, -3]"},
-      {RoundMode::TOWARDS_ZERO, "[3, 3, 3, 4, -3, -3, -3]"},
-      {RoundMode::TOWARDS_INFINITY, "[4, 4, 4, 5, -4, -4, -4]"},
-      {RoundMode::HALF_DOWN, "[3, 3, 4, 4, -3, -4, -4]"},
-      {RoundMode::HALF_UP, "[3, 4, 4, 5, -3, -3, -4]"},
-      {RoundMode::HALF_TOWARDS_ZERO, "[3, 3, 4, 4, -3, -3, -4]"},
-      {RoundMode::HALF_TOWARDS_INFINITY, "[3, 4, 4, 5, -3, -4, -4]"},
-      {RoundMode::HALF_TO_EVEN, "[3, 4, 4, 4, -3, -4, -4]"},
-      {RoundMode::HALF_TO_ODD, "[3, 3, 4, 5, -3, -3, -4]"},
-  }};
-  this->SetRoundNdigits(0);
-  for (const auto& pair : rmode_and_expected) {
-    this->SetRoundMode(pair.first);
-    this->AssertUnaryOp(Round, "[]", "[]");
-    this->AssertUnaryOp(Round, "[null, 0, Inf, -Inf, NaN, -NaN]",
-                        "[null, 0, Inf, -Inf, NaN, -NaN]");
-    this->AssertUnaryOp(Round, values, pair.second);
-  }
-
-  // Test different round N-digits for nearest rounding mode
-  values = "[320, 3.5, 3.075, 4.5, -3.212, -35.1234, -3.045]";
-  std::vector<std::pair<int64_t, std::string>> ndigits_and_expected{{
-      {-2, "[300, 0.0, 0.0, 0.0, -0.0, -0.0, -0.0]"},
-      {-1, "[320, 0.0, 0.0, 0.0, -0.0, -40, -0.0]"},
-      {0, "[320, 4, 3, 5, -3, -35, -3]"},
-      {1, "[320, 3.5, 3.1, 4.5, -3.2, -35.1, -3]"},
-      {2, "[320, 3.5, 3.08, 4.5, -3.21, -35.12, -3.05]"},
-  }};
-  this->SetRoundMode(RoundMode::HALF_TOWARDS_INFINITY);
-  for (const auto& pair : ndigits_and_expected) {
-    this->SetRoundNdigits(pair.first);
-    this->AssertUnaryOp(Round, values, pair.second);
-  }
-}
-
-TYPED_TEST_SUITE(TestUnaryRoundToMultipleIntegral, IntegralTypes);
-TYPED_TEST_SUITE(TestUnaryRoundToMultipleSigned, SignedIntegerTypes);
-TYPED_TEST_SUITE(TestUnaryRoundToMultipleUnsigned, UnsignedIntegerTypes);
-TYPED_TEST_SUITE(TestUnaryRoundToMultipleFloating, FloatingTypes);
-
-TYPED_TEST(TestUnaryRoundToMultipleSigned, RoundToMultiple) {
-  // Test different rounding modes for integer rounding
-  std::string values("[0, 1, -13, -50, 115]");
-  this->SetRoundMultiple(1);
-  for (const auto& round_mode : kRoundModes) {
-    this->SetRoundMode(round_mode);
-    this->AssertUnaryOp(RoundToMultiple, values, ArrayFromJSON(float64(), values));
-  }
-
-  // Test different round multiples for nearest rounding mode
-  std::vector<std::pair<double, std::string>> multiple_and_expected{{
-      {2, "[0.0, 2, -14, -50, 116]"},
-      {0.05, "[0.0, 1, -13, -50, 115]"},
-      {0.1, values},
-      {10, "[0.0, 0.0, -10, -50, 120]"},
-      {100, "[0.0, 0.0, -0.0, -100, 100]"},
-  }};
-  this->SetRoundMode(RoundMode::HALF_TOWARDS_INFINITY);
-  for (const auto& pair : multiple_and_expected) {
-    this->SetRoundMultiple(pair.first);
-    this->AssertUnaryOp(RoundToMultiple, values, ArrayFromJSON(float64(), pair.second));
-  }
-}
-
-TYPED_TEST(TestUnaryRoundToMultipleUnsigned, RoundToMultiple) {
-  // Test different rounding modes for integer rounding
-  std::string values("[0, 1, 13, 50, 115]");
-  this->SetRoundMultiple(1);
-  for (const auto& round_mode : kRoundModes) {
-    this->SetRoundMode(round_mode);
-    this->AssertUnaryOp(RoundToMultiple, values, ArrayFromJSON(float64(), values));
-  }
-
-  // Test different round multiples for nearest rounding mode
-  std::vector<std::pair<double, std::string>> multiple_and_expected{{
-      {0.05, "[0, 1, 13, 50, 115]"},
-      {0.1, values},
-      {2, "[0, 2, 14, 50, 116]"},
-      {10, "[0, 0, 10, 50, 120]"},
-      {100, "[0, 0, 0, 100, 100]"},
-  }};
-  this->SetRoundMode(RoundMode::HALF_TOWARDS_INFINITY);
-  for (const auto& pair : multiple_and_expected) {
-    this->SetRoundMultiple(pair.first);
-    this->AssertUnaryOp(RoundToMultiple, values, ArrayFromJSON(float64(), pair.second));
-  }
-}
-
-TYPED_TEST(TestUnaryRoundToMultipleFloating, RoundToMultiple) {
-  this->SetNansEqual(true);
-
-  // Test different rounding modes for integer rounding
-  std::string values("[3.2, 3.5, 3.7, 4.5, -3.2, -3.5, -3.7]");
-  std::vector<std::pair<RoundMode, std::string>> rmode_and_expected{{
-      {RoundMode::DOWN, "[3, 3, 3, 4, -4, -4, -4]"},
-      {RoundMode::UP, "[4, 4, 4, 5, -3, -3, -3]"},
-      {RoundMode::TOWARDS_ZERO, "[3, 3, 3, 4, -3, -3, -3]"},
-      {RoundMode::TOWARDS_INFINITY, "[4, 4, 4, 5, -4, -4, -4]"},
-      {RoundMode::HALF_DOWN, "[3, 3, 4, 4, -3, -4, -4]"},
-      {RoundMode::HALF_UP, "[3, 4, 4, 5, -3, -3, -4]"},
-      {RoundMode::HALF_TOWARDS_ZERO, "[3, 3, 4, 4, -3, -3, -4]"},
-      {RoundMode::HALF_TOWARDS_INFINITY, "[3, 4, 4, 5, -3, -4, -4]"},
-      {RoundMode::HALF_TO_EVEN, "[3, 4, 4, 4, -3, -4, -4]"},
-      {RoundMode::HALF_TO_ODD, "[3, 3, 4, 5, -3, -3, -4]"},
-  }};
-  this->SetRoundMultiple(1);
-  for (const auto& pair : rmode_and_expected) {
-    this->SetRoundMode(pair.first);
-    this->AssertUnaryOp(RoundToMultiple, "[]", "[]");
-    this->AssertUnaryOp(RoundToMultiple, "[null, 0, Inf, -Inf, NaN, -NaN]",
-                        "[null, 0, Inf, -Inf, NaN, -NaN]");
-    this->AssertUnaryOp(RoundToMultiple, values, pair.second);
-  }
-
-  // Test different round multiples for nearest rounding mode
-  values = "[320, 3.5, 3.075, 4.5, -3.212, -35.1234, -3.045]";
-  std::vector<std::pair<double, std::string>> multiple_and_expected{{
-      {0.05, "[320, 3.5, 3.1, 4.5, -3.2, -35.1, -3.05]"},
-      {0.1, "[320, 3.5, 3.1, 4.5, -3.2, -35.1, -3]"},
-      {2, "[320, 4, 4, 4, -4, -36, -4]"},
-      {10, "[320, 0.0, 0.0, 0.0, -0.0, -40, -0.0]"},
-      {100, "[300, 0.0, 0.0, 0.0, -0.0, -0.0, -0.0]"},
-  }};
-  this->SetRoundMode(RoundMode::HALF_TOWARDS_INFINITY);
-  for (const auto& pair : multiple_and_expected) {
-    this->SetRoundMultiple(pair.first);
-    this->AssertUnaryOp(RoundToMultiple, values, pair.second);
-  }
-
-  this->SetRoundMultiple(-2);
-  this->AssertUnaryOpRaises(RoundToMultiple, values,
-                            "Rounding multiple must be positive");
 }
 
 class TestBinaryArithmeticDecimal : public TestArithmeticDecimal {};
@@ -3504,156 +2797,5 @@ TYPED_TEST(TestUnaryArithmeticFloating, Sign) {
   this->AssertUnaryOp(sign, this->MakeScalar(max), this->MakeScalar(1));
 }
 
-TYPED_TEST(TestUnaryArithmeticSigned, Floor) {
-  auto floor = [](const Datum& arg, ArithmeticOptions, ExecContext* ctx) {
-    return Floor(arg, ctx);
-  };
-
-  this->AssertUnaryOp(floor, "[]", ArrayFromJSON(float64(), "[]"));
-  this->AssertUnaryOp(floor, "[null]", ArrayFromJSON(float64(), "[null]"));
-  this->AssertUnaryOp(floor, "[1, null, -10]",
-                      ArrayFromJSON(float64(), "[1, null, -10]"));
-  this->AssertUnaryOp(floor, "[0]", ArrayFromJSON(float64(), "[0]"));
-  this->AssertUnaryOp(floor, "[1, 10, 127]", ArrayFromJSON(float64(), "[1, 10, 127]"));
-  this->AssertUnaryOp(floor, "[-1, -10, -127]",
-                      ArrayFromJSON(float64(), "[-1, -10, -127]"));
-}
-
-TYPED_TEST(TestUnaryArithmeticUnsigned, Floor) {
-  auto floor = [](const Datum& arg, ArithmeticOptions, ExecContext* ctx) {
-    return Floor(arg, ctx);
-  };
-
-  this->AssertUnaryOp(floor, "[]", ArrayFromJSON(float64(), "[]"));
-  this->AssertUnaryOp(floor, "[null]", ArrayFromJSON(float64(), "[null]"));
-  this->AssertUnaryOp(floor, "[1, null, 10]", ArrayFromJSON(float64(), "[1, null, 10]"));
-  this->AssertUnaryOp(floor, "[0]", ArrayFromJSON(float64(), "[0]"));
-  this->AssertUnaryOp(floor, "[1, 10, 127]", ArrayFromJSON(float64(), "[1, 10, 127]"));
-}
-
-TYPED_TEST(TestUnaryArithmeticFloating, Floor) {
-  using CType = typename TestFixture::CType;
-  auto min = std::numeric_limits<CType>::lowest();
-  auto max = std::numeric_limits<CType>::max();
-
-  this->SetNansEqual(true);
-
-  auto floor = [](const Datum& arg, ArithmeticOptions, ExecContext* ctx) {
-    return Floor(arg, ctx);
-  };
-
-  this->AssertUnaryOp(floor, "[]", "[]");
-  this->AssertUnaryOp(floor, "[null]", "[null]");
-  this->AssertUnaryOp(floor, "[1.3, null, -10.80]", "[1, null, -11]");
-  // XXX Python uses math.floor(-0.0) == 0.0, but std::floor() keeps the sign
-  this->AssertUnaryOp(floor, "[0.0, -0.0]", "[0.0, -0.0]");
-  this->AssertUnaryOp(floor, "[1.3, 10.80, 12748.001]", "[1, 10, 12748]");
-  this->AssertUnaryOp(floor, "[-1.3, -10.80, -12748.001]", "[-2, -11, -12749]");
-  this->AssertUnaryOp(floor, "[Inf, -Inf]", "[Inf, -Inf]");
-  this->AssertUnaryOp(floor, "[NaN]", "[NaN]");
-  this->AssertUnaryOp(floor, this->MakeScalar(min), this->MakeScalar(min));
-  this->AssertUnaryOp(floor, this->MakeScalar(max), this->MakeScalar(max));
-}
-
-TYPED_TEST(TestUnaryArithmeticSigned, Ceil) {
-  auto ceil = [](const Datum& arg, ArithmeticOptions, ExecContext* ctx) {
-    return Ceil(arg, ctx);
-  };
-
-  this->AssertUnaryOp(ceil, "[]", ArrayFromJSON(float64(), "[]"));
-  this->AssertUnaryOp(ceil, "[null]", ArrayFromJSON(float64(), "[null]"));
-  this->AssertUnaryOp(ceil, "[1, null, -10]", ArrayFromJSON(float64(), "[1, null, -10]"));
-  this->AssertUnaryOp(ceil, "[0]", ArrayFromJSON(float64(), "[0]"));
-  this->AssertUnaryOp(ceil, "[1, 10, 127]", ArrayFromJSON(float64(), "[1, 10, 127]"));
-  this->AssertUnaryOp(ceil, "[-1, -10, -127]",
-                      ArrayFromJSON(float64(), "[-1, -10, -127]"));
-}
-
-TYPED_TEST(TestUnaryArithmeticUnsigned, Ceil) {
-  auto ceil = [](const Datum& arg, ArithmeticOptions, ExecContext* ctx) {
-    return Ceil(arg, ctx);
-  };
-
-  this->AssertUnaryOp(ceil, "[]", ArrayFromJSON(float64(), "[]"));
-  this->AssertUnaryOp(ceil, "[null]", ArrayFromJSON(float64(), "[null]"));
-  this->AssertUnaryOp(ceil, "[1, null, 10]", ArrayFromJSON(float64(), "[1, null, 10]"));
-  this->AssertUnaryOp(ceil, "[0]", ArrayFromJSON(float64(), "[0]"));
-  this->AssertUnaryOp(ceil, "[1, 10, 127]", ArrayFromJSON(float64(), "[1, 10, 127]"));
-}
-
-TYPED_TEST(TestUnaryArithmeticFloating, Ceil) {
-  using CType = typename TestFixture::CType;
-  auto min = std::numeric_limits<CType>::lowest();
-  auto max = std::numeric_limits<CType>::max();
-
-  this->SetNansEqual(true);
-
-  auto ceil = [](const Datum& arg, ArithmeticOptions, ExecContext* ctx) {
-    return Ceil(arg, ctx);
-  };
-
-  this->AssertUnaryOp(ceil, "[]", "[]");
-  this->AssertUnaryOp(ceil, "[null]", "[null]");
-  this->AssertUnaryOp(ceil, "[1.3, null, -10.80]", "[2, null, -10]");
-  // XXX same comment as Floor above
-  this->AssertUnaryOp(ceil, "[0.0, -0.0]", "[0.0, -0.0]");
-  this->AssertUnaryOp(ceil, "[1.3, 10.80, 12748.001]", "[2, 11, 12749]");
-  this->AssertUnaryOp(ceil, "[-1.3, -10.80, -12748.001]", "[-1, -10, -12748]");
-  this->AssertUnaryOp(ceil, "[Inf, -Inf]", "[Inf, -Inf]");
-  this->AssertUnaryOp(ceil, "[NaN]", "[NaN]");
-  this->AssertUnaryOp(ceil, this->MakeScalar(min), this->MakeScalar(min));
-  this->AssertUnaryOp(ceil, this->MakeScalar(max), this->MakeScalar(max));
-}
-
-TYPED_TEST(TestUnaryArithmeticSigned, Trunc) {
-  auto trunc = [](const Datum& arg, ArithmeticOptions, ExecContext* ctx) {
-    return Trunc(arg, ctx);
-  };
-
-  this->AssertUnaryOp(trunc, "[]", ArrayFromJSON(float64(), "[]"));
-  this->AssertUnaryOp(trunc, "[null]", ArrayFromJSON(float64(), "[null]"));
-  this->AssertUnaryOp(trunc, "[1, null, -10]",
-                      ArrayFromJSON(float64(), "[1, null, -10]"));
-  this->AssertUnaryOp(trunc, "[0]", ArrayFromJSON(float64(), "[0]"));
-  this->AssertUnaryOp(trunc, "[1, 10, 127]", ArrayFromJSON(float64(), "[1, 10, 127]"));
-  this->AssertUnaryOp(trunc, "[-1, -10, -127]",
-                      ArrayFromJSON(float64(), "[-1, -10, -127]"));
-}
-
-TYPED_TEST(TestUnaryArithmeticUnsigned, Trunc) {
-  auto trunc = [](const Datum& arg, ArithmeticOptions, ExecContext* ctx) {
-    return Trunc(arg, ctx);
-  };
-
-  this->AssertUnaryOp(trunc, "[]", ArrayFromJSON(float64(), "[]"));
-  this->AssertUnaryOp(trunc, "[null]", ArrayFromJSON(float64(), "[null]"));
-  this->AssertUnaryOp(trunc, "[1, null, 10]", ArrayFromJSON(float64(), "[1, null, 10]"));
-  this->AssertUnaryOp(trunc, "[0]", ArrayFromJSON(float64(), "[0]"));
-  this->AssertUnaryOp(trunc, "[1, 10, 127]", ArrayFromJSON(float64(), "[1, 10, 127]"));
-}
-
-TYPED_TEST(TestUnaryArithmeticFloating, Trunc) {
-  using CType = typename TestFixture::CType;
-  auto min = std::numeric_limits<CType>::lowest();
-  auto max = std::numeric_limits<CType>::max();
-
-  this->SetNansEqual(true);
-
-  auto trunc = [](const Datum& arg, ArithmeticOptions, ExecContext* ctx) {
-    return Trunc(arg, ctx);
-  };
-
-  this->AssertUnaryOp(trunc, "[]", "[]");
-  this->AssertUnaryOp(trunc, "[null]", "[null]");
-  this->AssertUnaryOp(trunc, "[1.3, null, -10.80]", "[1, null, -10]");
-  // XXX same comment as Floor above
-  this->AssertUnaryOp(trunc, "[0.0, -0.0]", "[0.0, -0.0]");
-  this->AssertUnaryOp(trunc, "[1.3, 10.80, 12748.001]", "[1, 10, 12748]");
-  this->AssertUnaryOp(trunc, "[-1.3, -10.80, -12748.001]", "[-1, -10, -12748]");
-  this->AssertUnaryOp(trunc, "[Inf, -Inf]", "[Inf, -Inf]");
-  this->AssertUnaryOp(trunc, "[NaN]", "[NaN]");
-  this->AssertUnaryOp(trunc, this->MakeScalar(min), this->MakeScalar(min));
-  this->AssertUnaryOp(trunc, this->MakeScalar(max), this->MakeScalar(max));
-}
 }  // namespace compute
 }  // namespace arrow
